[parser] move primitive value parsing
to module primitive_value
